fileloader/pyproject.toml
[project]
name = "fileloader"
version = "0.1.0"
description = "Bulk conversion from document files to JSONL dataset."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fsspec>=2024.6.1",
    "openpyxl>=3.1.5",
    "pymupdf>=1.24.9",
    "python-pptx>=1.0.2",
    "pydantic>=2.8.2",
    "tqdm>=4.66.5",
    "pandas>=2.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
